+++ internal/server/v1/info/golden/bulk_place_info/bulk_place_info.json
