+++ internal/server/v1/info/golden/bulk_place_info_test.go
