+++ internal/server/v1/info/golden/place_info_test.go
