+++ internal/server/v1/info/place.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package place
+package info
 
 import (
 	"context"
